+++ CHANGES.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+* replace custom `BoundingBox` with `BoundingBox` from openEO pg_schema to fix type validation errors
+
 ## [0.6.0] (2025-12-04)
 
 ### Changed